Changed ap storage size from i8 to i32. The recent geomagnetic storm reached values >128. Also added launch.json VSCode config for easy debugging.

+++ src/kp_data.rs
@@ -87,7 +87,7 @@ impl FromStr for KpFile {
 pub struct Entry {
     pub date: NaiveDateTime,
     pub kp: f32,
-    pub ap: i8,
+    pub ap: i32,
     pub d: i8
 }
